--- go/vt/vtgate/planbuilder/operators/compact.go
@@ -1,115 +1,0 @@
-/*
-Copyright 2022 The Vitess Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package operators
-
-import (
-	"vitess.io/vitess/go/vt/sqlparser"
-	"vitess.io/vitess/go/vt/vtgate/planbuilder/operators/ops"
-	"vitess.io/vitess/go/vt/vtgate/planbuilder/operators/rewrite"
-	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
-)
-
-// compact will optimise the operator tree into a smaller but equivalent version
-func compact(ctx *plancontext.PlanningContext, op ops.Operator) (ops.Operator, error) {
-	newOp, err := rewrite.BottomUp(op, func(op ops.Operator) (ops.Operator, rewrite.TreeIdentity, error) {
-		newOp, ok := op.(compactable)
-		if !ok {
-			return op, rewrite.SameTree, nil
-		}
-		return newOp.compact(ctx)
-	})
-	return newOp, err
-}
-
-func (f *Filter) compact(*plancontext.PlanningContext) (ops.Operator, rewrite.TreeIdentity, error) {
-	if len(f.Predicates) == 0 {
-		return f.Source, rewrite.NewTree, nil
-	}
-
-	other, isFilter := f.Source.(*Filter)
-	if !isFilter {
-		return f, rewrite.SameTree, nil
-	}
-	f.Source = other.Source
-	f.Predicates = append(f.Predicates, other.Predicates...)
-	return f, rewrite.NewTree, nil
-}
-
-func (u *Union) compact(*plancontext.PlanningContext) (ops.Operator, rewrite.TreeIdentity, error) {
-	var newSources []ops.Operator
-	anythingChanged := false
-	for _, source := range u.Sources {
-		var other *Union
-		horizon, ok := source.(*Horizon)
-		if ok {
-			union, ok := horizon.Source.(*Union)
-			if ok {
-				other = union
-			}
-		}
-		if other == nil {
-			newSources = append(newSources, source)
-			continue
-		}
-		anythingChanged = true
-		switch {
-		case len(other.Ordering) == 0 && !other.Distinct:
-			fallthrough
-		case u.Distinct:
-			// if the current UNION is a DISTINCT, we can safely ignore everything from children UNIONs, except LIMIT
-			newSources = append(newSources, other.Sources...)
-
-		default:
-			newSources = append(newSources, other)
-		}
-	}
-	if anythingChanged {
-		u.Sources = newSources
-	}
-	identity := rewrite.SameTree
-	if anythingChanged {
-		identity = rewrite.NewTree
-	}
-
-	return u, identity, nil
-}
-
-func (j *Join) compact(ctx *plancontext.PlanningContext) (ops.Operator, rewrite.TreeIdentity, error) {
-	if j.LeftJoin {
-		// we can't merge outer joins into a single QG
-		return j, rewrite.SameTree, nil
-	}
-
-	lqg, lok := j.LHS.(*QueryGraph)
-	rqg, rok := j.RHS.(*QueryGraph)
-	if !lok || !rok {
-		return j, rewrite.SameTree, nil
-	}
-
-	newOp := &QueryGraph{
-		Tables:     append(lqg.Tables, rqg.Tables...),
-		innerJoins: append(lqg.innerJoins, rqg.innerJoins...),
-		NoDeps:     sqlparser.AndExpressions(lqg.NoDeps, rqg.NoDeps),
-	}
-	if j.Predicate != nil {
-		err := newOp.collectPredicate(ctx, j.Predicate)
-		if err != nil {
-			return nil, rewrite.SameTree, err
-		}
-	}
-	return newOp, rewrite.NewTree, nil
-}
