+++ go/vt/vtgate/planbuilder/operators/join.go
@@ -19,8 +19,8 @@ package operators
 import (
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/vtgate/planbuilder/operators/ops"
+	"vitess.io/vitess/go/vt/vtgate/planbuilder/operators/rewrite"
 	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
-	"vitess.io/vitess/go/vt/vtgate/semantics"
 )
 
 // Join represents a join. If we have a predicate, this is an inner join. If no predicate exists, it is a cross join
@@ -36,7 +36,6 @@ var _ ops.Operator = (*Join)(nil)
 
 // Clone implements the Operator interface
 func (j *Join) Clone(inputs []ops.Operator) ops.Operator {
-	checkSize(inputs, 2)
 	clone := *j
 	clone.LHS = inputs[0]
 	clone.RHS = inputs[1]
@@ -51,6 +50,32 @@ func (j *Join) Clone(inputs []ops.Operator) ops.Operator {
 // Inputs implements the Operator interface
 func (j *Join) Inputs() []ops.Operator {
 	return []ops.Operator{j.LHS, j.RHS}
+}
+
+func (j *Join) Compact(ctx *plancontext.PlanningContext) (ops.Operator, rewrite.TreeIdentity, error) {
+	if j.LeftJoin {
+		// we can't merge outer joins into a single QG
+		return j, rewrite.SameTree, nil
+	}
+
+	lqg, lok := j.LHS.(*QueryGraph)
+	rqg, rok := j.RHS.(*QueryGraph)
+	if !lok || !rok {
+		return j, rewrite.SameTree, nil
+	}
+
+	newOp := &QueryGraph{
+		Tables:     append(lqg.Tables, rqg.Tables...),
+		innerJoins: append(lqg.innerJoins, rqg.innerJoins...),
+		NoDeps:     sqlparser.AndExpressions(lqg.NoDeps, rqg.NoDeps),
+	}
+	if j.Predicate != nil {
+		err := newOp.collectPredicate(ctx, j.Predicate)
+		if err != nil {
+			return nil, rewrite.SameTree, err
+		}
+	}
+	return newOp, rewrite.NewTree, nil
 }
 
 func createOuterJoin(tableExpr *sqlparser.JoinTableExpr, lhs, rhs ops.Operator) (ops.Operator, error) {
@@ -88,40 +113,36 @@ func createInnerJoin(ctx *plancontext.PlanningContext, tableExpr *sqlparser.Join
 }
 
 func (j *Join) AddPredicate(ctx *plancontext.PlanningContext, expr sqlparser.Expr) (ops.Operator, error) {
-	return addPredicate(j, ctx, expr, false)
+	return AddPredicate(j, ctx, expr, false, newFilter)
 }
 
-var _ joinOperator = (*Join)(nil)
+var _ JoinOp = (*Join)(nil)
 
-func (j *Join) tableID() semantics.TableSet {
-	return TableID(j)
-}
-
-func (j *Join) getLHS() ops.Operator {
+func (j *Join) GetLHS() ops.Operator {
 	return j.LHS
 }
 
-func (j *Join) getRHS() ops.Operator {
+func (j *Join) GetRHS() ops.Operator {
 	return j.RHS
 }
 
-func (j *Join) setLHS(operator ops.Operator) {
+func (j *Join) SetLHS(operator ops.Operator) {
 	j.LHS = operator
 }
 
-func (j *Join) setRHS(operator ops.Operator) {
+func (j *Join) SetRHS(operator ops.Operator) {
 	j.RHS = operator
 }
 
-func (j *Join) makeInner() {
+func (j *Join) MakeInner() {
 	j.LeftJoin = false
 }
 
-func (j *Join) isInner() bool {
+func (j *Join) IsInner() bool {
 	return !j.LeftJoin
 }
 
-func (j *Join) addJoinPredicate(_ *plancontext.PlanningContext, expr sqlparser.Expr) error {
+func (j *Join) AddJoinPredicate(_ *plancontext.PlanningContext, expr sqlparser.Expr) error {
 	j.Predicate = sqlparser.AndExpressions(j.Predicate, expr)
 	return nil
 }
