+++ go/vt/vtgate/planbuilder/operators/operator_test.go
@@ -105,7 +105,7 @@ func TestOperator(t *testing.T) {
 			ctx := plancontext.NewPlanningContext(nil, semTable, nil, 0)
 			optree, err := createLogicalOperatorFromAST(ctx, stmt)
 			require.NoError(t, err)
-			optree, err = compact(ctx, optree)
+			optree, err = Compact(ctx, optree)
 			require.NoError(t, err)
 			output := testString(optree)
 			assert.Equal(t, tc.expected, output)
